changed octree implementation to use ull

+++ src/phys/slicing.cpp
@@ -31,7 +31,7 @@ engine::Scene *tick(engine::Scene *e)
         collider *c = go -> getComponent<collider>();
         aabb caabb = c -> to_aabb();
         BOOST_LOG_TRIVIAL(trace) << "Checking for colisions";
-        std::set<int> collisions;
+        std::set<unsigned long long> collisions;
         collisions = e -> ctree.get_collisions(caabb);
         collisions.erase(i);
         BOOST_LOG_TRIVIAL(trace) << "Handling collision";
